Split SQLiteStorageAdapter tests to reduce load
